requirements.txt
Django==2.2
djangorestframework==3.13.1
esdk-obs-python==3.20.11
requests==2.24.0
